return a task view instead of a task
